requirements.txt
paho-mqtt
pandas
numpy
matplotlib
scikit-learn
